+++ src/jvmMain/kotlin/com/sunnychung/application/multiplatform/hellohttp/ux/AppTextFieldWithVariables.kt
@@ -17,6 +17,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
+import com.sunnychung.application.multiplatform.hellohttp.util.letIf
 import com.sunnychung.application.multiplatform.hellohttp.util.log
 import com.sunnychung.application.multiplatform.hellohttp.ux.AppUX.ENV_VAR_VALUE_MAX_DISPLAY_LENGTH
 import com.sunnychung.application.multiplatform.hellohttp.ux.local.LocalColor
@@ -124,8 +125,8 @@ fun AppTextFieldWithVariables(
             singleLine = singleLine,
             colors = colors,
             contentPadding = contentPadding,
-            onPointerEvent = if (isSupportVariables) {
-                { event, tag ->
+            onPointerEvent = letIf(isSupportVariables) {
+                { event, _, tag ->
                     log.v { "onPointerEventOnAnnotatedTag $tag $event" }
                     mouseHoverVariable =
                         if (tag?.startsWith(EnvironmentVariableIncrementalTransformation.TAG_PREFIX) == true) {
@@ -134,8 +135,6 @@ fun AppTextFieldWithVariables(
                             null
                         }
                 }
-            } else {
-                null
             },
             interactionSource = interactionSource,
             onFinishInit = onFinishInit,